requirements.txt
discord.py==2.5.2
Flask==3.1.1
python-dotenv==1.1.1
numpy==2.3.1
Pyrebase4==4.8.0
urllib3==1.26.20
setuptools==80.9.0